feat: implement full-stack trend intelligence system with data pipelines, ML engines, and frontend dashboard

+++ backend/app/schemas/search_schema.py
@@ -1,8 +1,9 @@
 # Define Pydantic request/response schemas for data validation
 from pydantic import BaseModel
 
+
 # Schema defining the structure and data types of a search response
 class SearchResponse(BaseModel):
-    query: str         # The original search term typed by the user
-    trend_score: int   # The calculated popularity/trend score for the query
-    message: str       # A status message indicating the processing result
+    query: str          # The original search term typed by the user
+    trend_score: float  # The ML-computed popularity/trend score for the query
+    message: str        # A status message indicating the processing result
